+++ CHANGELOG.md
@@ -1,4 +1,16 @@
 
+## 0.0.5
+### Added
+- Get-WBWorkspace cmdlet
+- New-WBWorkspace cmdlet
+- Switch-WBWorkspace cmdlet
+- Workspace templates
+### Changed
+- Minor refactoring
+### Removed
+- $Workbench variable
+- Edit-WBConfiguration aliases
+
 ## 0.0.4
 ### Added
 - Edit-WBConfiguration cmdlet